+++ vehicle-python/src/vehicle_lang/list/__init__.py
@@ -8,9 +8,9 @@ from typing_extensions import TypeAlias
 from typing_extensions import TypeVar as TypingTypeVar
 
 from .. import session
-from ..error import VehicleError
-from ..loss._ast._decode import JsonValue, decode
-from ..loss._ast._nodes import Provenance
+from ..error import VehicleInternalError
+from .._ast._decode import JsonValue, decode
+from .._ast._nodes import Provenance
 
 Quantifier: TypeAlias = str
 
@@ -109,18 +109,14 @@ def list(specification: str | Path) -> List[ListableEntity]:
     args = ["list", "--specification", str(specification), "--json"]
 
     # Call Vehicle
-    exc, out, err, _ = session.check_output(args)
-
-    # Check for errors
-    if exc != 0:
-        raise VehicleError(f"{err}")
-    elif not out:
+    out = session.execute_command(args)
+    if not out:
         return []
 
     try:
         return _decode_listable_entities(json.loads(out))
     except Exception as exc:
-        raise VehicleError(str(exc)) from exc
+        raise VehicleInternalError(str(exc)) from exc
 
 
 __all__ = [