+++ vehicle/src/Vehicle/Resource.hs
@@ -2,7 +2,7 @@
 
 module Vehicle.Resource where
 
-import Control.Exception (IOException, catch)
+import Control.Exception (IOException, catch, try)
 import Control.Monad (forM)
 import Control.Monad.IO.Class (MonadIO, liftIO)
 import Data.Aeson (FromJSON, ToJSON)
@@ -14,7 +14,7 @@ import Data.Map qualified as Map
 import Data.Text (Text)
 import GHC.Generics (Generic)
 import Vehicle.Prelude (Name)
-import Vehicle.Prelude.IO (fatalError)
+import Vehicle.Prelude.IO (fatalError, MonadStdIO)
 import Vehicle.Prelude.Prettyprinter
 
 --------------------------------------------------------------------------------
@@ -136,29 +136,30 @@ hashFileContents filePath = do
   fileContents <- liftIO $ ByteString.readFile filePath
   return $ hash fileContents
 
-generateResourceIntegrityInfo :: (MonadIO m) => (Name, FilePath) -> m ResourceIntegrityInfo
+generateResourceIntegrityInfo :: (MonadStdIO m) => (Name, FilePath) -> m ResourceIntegrityInfo
 generateResourceIntegrityInfo (name, filePath) = do
-  fileHash <-
+  errorOrfileHash <-
     liftIO $
-      catch @IOException
+      try @IOException
         (hashFileContents filePath)
-        ( \e ->
+        
+  case errorOrfileHash of
+    Left err -> do
             fatalError $
               "Error occured while reading"
                 <+> quotePretty filePath
                 <> ":"
                 <> line
-                <> indent 2 (pretty (show e))
-        )
+                <> indent 2 (pretty (show err))
+    Right fileHash -> do
+      return $
+        ResourceIntegrityInfo
+          { name = name,
+            filePath = filePath,
+            fileHash = fileHash
+          }
 
-  return $
-    ResourceIntegrityInfo
-      { name = name,
-        filePath = filePath,
-        fileHash = fileHash
-      }
-
-generateResourcesIntegrityInfo :: (MonadIO m) => Resources -> m ResourcesIntegrityInfo
+generateResourcesIntegrityInfo :: (MonadStdIO m) => Resources -> m ResourcesIntegrityInfo
 generateResourcesIntegrityInfo Resources {..} = do
   specificationSummary <- generateResourceIntegrityInfo ("specification", specification)
   networkSummaries <- forM (assocs networks) generateResourceIntegrityInfo